third_party/WebKit/LayoutTests/fast/canvas/canvas-strokeText-strokes-shadow.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function(t) {
    var ctx = document.createElement('canvas').getContext('2d');
    
    ctx.fillStyle = 'green';
    ctx.fillRect(0, 0, ctx.canvas.width, ctx.canvas.height);
    
    // Stroke an 'I' with its shadow in the upper left corner.
    ctx.strokeStyle = 'white';
    ctx.lineWidth = 2;
    ctx.shadowColor = 'red';
    ctx.shadowOffsetX = -15;
    ctx.shadowOffsetY = 0;
    ctx.font = '128px sans-serif';
    ctx.strokeText("I", 0, 50);
    
    assert_array_equals(ctx.getImageData(0, 0, 1, 1).data, [0, 128, 0, 255]);
}, "Test that strokeText() doesn't produce a filled shadow.");
</script>
</body>
